+++ lib/zaikio/procurement/article.rb
@@ -1,13 +1,16 @@
+require_relative "concern/article_type_scoped"
+
 module Zaikio
   module Procurement
     class Article < Base
+      include ArticleTypeScoped
       include_root_in_json :article
 
       # Class methods
       class << self
         # Spyke URI override
         def uri
-          Zaikio::Procurement.configuration.flavor == :supplier ? "substrate/articles(/:id)" : "articles(/:id)"
+          Zaikio::Procurement.configuration.flavor == :supplier ? ":type/articles(/:id)" : "articles(/:id)"
         end
 
         def list_by_article_type_or_supplier_slug(type_or_slug)
@@ -24,8 +27,8 @@ module Zaikio
                  :kind, :supplier_id, :created_at, :updated_at
 
       # Associations
-      has_one :supplier,  class_name: "Zaikio::Procurement::Supplier",
-                          uri: nil
+      has_one :supplier, class_name: "Zaikio::Procurement::Supplier",
+                         uri: nil
       # Manually build variants association to work for consumers and suppliers
       def variants
         self.class.request(:get, "#{uri}/variants").data.collect { |v| Zaikio::Procurement::Variant.new(v) }